+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Validate mode solver object for large number of grid points on the modal plane.
+
 ## [2.9.0] - 2025-08-04
 
 ### Added